The basic particle system is now functional. The billboard group now uses vertex and index buffers for data storage. Lots of work still needs to be done to expose configuration options for particle systems. Particle systems can now use an animation instead of a single texture. Added a couple of really simply particle effects for changing color. Particle effects still needs some thought as to how to apply them and make them generic. Optimizations need to be made because adding a single effect right now halves the FPS on an older graphics card.

+++ branches/Engine_0_2/src/graphics/IndexData.h
@@ -29,16 +29,16 @@ class IndexData
 public:
 	IndexData();
 	IndexData(uint32_t numIndices);
-	IndexData(float* indices, uint32_t numIndices);
+	IndexData(uint16_t* indices, uint32_t numIndices);
 
-	void AddIndex(float index);
-	void AddIndices(float* indices, uint32_t numIndices);
+	void AddIndex(uint16_t index);
+	void AddIndices(uint16_t* indices, uint32_t numIndices);
 	void Clear();
-	float* GetIndices();
+	uint16_t* GetIndices();
 	uint32_t GetNumIndices() const;
 	
 private:
-	typedef std::vector<float> IndexContainer;
+	typedef std::vector<uint16_t> IndexContainer;
 	IndexContainer m_indices;
 };
